requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
numpy==2.3.5
scipy==1.17.1
scikit-learn==1.8.0
pandas==3.0.1
matplotlib==3.10.8
h5py==3.15.1
PyYAML==6.0.3
tqdm==4.67.3
joblib==1.5.3
torch==2.10.0+cpu
torchvision==0.25.0+cpu
torchaudio==2.10.0+cpu
efficient-kan @ git+https://github.com/Blealtan/efficient-kan.git@7b6ce1c87f18c8bc90c208f6b494042344216b11
